Move alarm_system to models
